requirements.txt
python-binance==1.0.12
aiohttp==3.8.4
python-dotenv==1.0.0
pytest==8.2.0
pytest-asyncio==0.23.6
hypothesis==6.112.1
